Another refactoring after deeper thought on the job scheduler
The job scheduler does not work 1-to-1 with the actions and executions, as it has a different view of how things should work.  This change allows for simplifying the design, but puts more work on the builder.  However, it should allow for the builder design to, in the end, work more naturally with setting up parallel states.